+++ packages/hy_fluttertoast/CHANGELOG.md
@@ -1,6 +1,14 @@
-## [0.0.2] - 2020-12-24 15:33:57
+## [0.0.4] - 2020-12-24 16:35:13
 
-* `IToastService.show` add `timeInSec` parameter.
+* `ToastService.show` 增加默认配置类实现
+
+## [0.0.3] - 2020-12-24 15:33:57
+
+* `ToastService.show` add `timeInSec` parameter.
+
+## [0.0.2] - 2020-12-21 14:33:33
+
+* `hy_fluttertoast`中的`fluttertoast`升级到v7.1.6
 
 ## [0.0.1] - 2020-10-28 17:23:01